Move gapclosing C++ code to its own source file
